Geometric baselines at 1.10; K scales with attacker level; decimal damage and defense
Every item and mob baseline is its level-1 value times
levelGrowth^(level-1), parameterized in rules.js. K grows with the
attacker's level at the same rate (kGrowth 1.0) so a tier of armor
reduces the same share at every level and fight length no longer drifts.
Weapon damage and armor defense are float64 so rounding does not distort
low levels. Balance dummies extended to level 25. Measurements and the
history of the choice recorded in RULES.md 4.5.

+++ internal/item/proto.go
@@ -102,7 +102,7 @@ type Proto struct {
 // Spread the fraction it varies by, Speed the swings per round, Verb the
 // word combat messages use ("slash", "bite").
 type WeaponSpec struct {
-	Damage int     `yaml:"damage" json:"damage"`
+	Damage float64 `yaml:"damage" json:"damage"`
 	Spread float64 `yaml:"spread" json:"spread"`
 	Speed  float64 `yaml:"speed" json:"speed"`
 	Hands  int     `yaml:"hands" json:"hands"`
@@ -112,7 +112,7 @@ type WeaponSpec struct {
 
 // ArmorSpec is rule input for armor.
 type ArmorSpec struct {
-	Defense int     `yaml:"defense" json:"defense"`
+	Defense float64 `yaml:"defense" json:"defense"`
 	Spread  float64 `yaml:"spread" json:"spread"`
 }
